GLFW, WinWindow, window events

+++ Deyo/src/Deyo/Events/ApplicationEvent.h
@@ -35,17 +35,17 @@ namespace Deyo
 	class WindowResizeEvent : public Event
 	{
 	public:
-		WindowResizeEvent(uint32_t width, uint32_t height) : m_Width(width), m_Height(height) { };
+		WindowResizeEvent(uint16_t width, uint16_t height) : m_Width(width), m_Height(height) { };
 
-		inline uint32_t GetWidth() { return m_Width; }
-		inline uint32_t GetHeight() { return m_Height; }
+		inline uint16_t GetWidth() { return m_Width; }
+		inline uint16_t GetHeight() { return m_Height; }
 
 		EVENT_TYPE(EventType::WindowResize)
 			EVENT_CATEGORY(EventCategoryApplication)
 
 	private:
-		uint32_t m_Width;
-		uint32_t m_Height;
+		uint16_t m_Width;
+		uint16_t m_Height;
 	};
 
 	class WindowMoveEvent : public Event
